TST: unit test for custom routers
