+++ bluesky_httpserver/tests/test_router_general_fs.py
@@ -25,8 +25,6 @@ from bluesky_httpserver.tests.conftest import (  # noqa F401
     wait_for_manager_state_idle,
 )
 
-from bluesky_queueserver import generate_zmq_keys
-
 
 # Plans used in most of the tests: '_plan1' and '_plan2' are quickly executed '_plan3' runs for 5 seconds.
 _plan1 = {"name": "count", "args": [["det1", "det2"]], "item_type": "plan"}
@@ -301,125 +299,3 @@ def test_http_server_queue_upload_spreasheet_5(re_manager, fastapi_server_fs, tm
     resp2 = request_to_json("get", "/status")
     assert resp2["items_in_queue"] == 0
     assert resp2["items_in_history"] == 0
-
-
-# fmt: off
-@pytest.mark.parametrize("test_mode", ["none", "ev"])
-# fmt: on
-def test_http_server_secure_1(monkeypatch, re_manager_cmd, fastapi_server_fs, test_mode):  # noqa: F811
-    """
-    Test operation of HTTP server with enabled encryption. Security of HTTP server can be enabled
-    only by setting the environment variable to the value of the public key.
-    """
-    public_key, private_key = generate_zmq_keys()
-
-    if test_mode == "none":
-        # No encryption
-        pass
-    elif test_mode == "ev":
-        # Set server private key using environment variable
-        monkeypatch.setenv("QSERVER_ZMQ_PRIVATE_KEY_FOR_SERVER", private_key)  # RE Manager
-        monkeypatch.setenv("QSERVER_ZMQ_PUBLIC_KEY", public_key)  # HTTP server
-        set_qserver_zmq_public_key(monkeypatch, server_public_key=public_key)  # For test functions
-    else:
-        raise RuntimeError(f"Unrecognized test mode '{test_mode}'")
-
-    fastapi_server_fs()
-    re_manager_cmd([])
-
-    resp1 = request_to_json("post", "/queue/item/add", json={"item": _plan1})
-    assert resp1["success"] is True, str(resp1)
-
-    resp2 = request_to_json("post", "/queue/item/add", json={"item": _plan2})
-    assert resp2["success"] is True, str(resp2)
-
-    resp3 = request_to_json("get", "/plans/allowed")
-    assert isinstance(resp3["plans_allowed"], dict)
-    assert len(resp3["plans_allowed"]) > 0
-    resp4 = request_to_json("get", "/devices/allowed")
-    assert isinstance(resp4["devices_allowed"], dict)
-    assert len(resp4["devices_allowed"]) > 0
-
-    resp5 = request_to_json("post", "/environment/open")
-    assert resp5["success"] is True
-    assert wait_for_environment_to_be_created(10)
-
-    resp6 = request_to_json("get", "/status")
-    assert resp6["items_in_queue"] == 2
-    assert resp6["items_in_history"] == 0
-
-    resp7 = request_to_json("post", "/queue/start")
-    assert resp7["success"] is True
-
-    wait_for_queue_execution_to_complete(20)
-
-    resp8 = request_to_json("get", "/status")
-    assert resp8["items_in_queue"] == 0
-    assert resp8["items_in_history"] == 2
-
-    # Close the environment
-    resp9 = request_to_json("post", "/environment/close")
-    assert resp9 == {"success": True, "msg": ""}
-
-    wait_for_manager_state_idle(10)
-
-
-def test_http_server_set_zmq_address_1(monkeypatch, re_manager_cmd, fastapi_server_fs):  # noqa: F811
-    """
-    Test if ZMQ address of RE Manager is passed to the HTTP server using 'QSERVER_ZMQ_ADDRESS_CONTROL'
-    environment variable. Start RE Manager and HTTP server with ZMQ address for control communication
-    channel different from default address, add and execute a plan.
-    """
-
-    # Change ZMQ address to use port 60616 instead of the default port 60615.
-    zmq_control_address_server = "tcp://*:60616"
-    zmq_info_address_server = "tcp://*:60617"
-    zmq_control_address = "tcp://localhost:60616"
-    zmq_info_address = "tcp://localhost:60617"
-    monkeypatch.setenv("QSERVER_ZMQ_CONTROL_ADDRESS", zmq_control_address)
-    monkeypatch.setenv("QSERVER_ZMQ_INFO_ADDRESS", zmq_info_address)
-    fastapi_server_fs()
-
-    set_qserver_zmq_address(monkeypatch, zmq_server_address=zmq_control_address)
-    re_manager_cmd(
-        [
-            f"--zmq-control-addr={zmq_control_address_server}",
-            f"--zmq-info-addr={zmq_info_address_server}",
-            "--zmq-publish-console=ON",
-        ]
-    )
-
-    # Now execute a plan to make sure everything works as expected
-    resp1 = request_to_json("post", "/queue/item/add", json={"item": _plan1})
-    assert resp1["success"] is True, str(resp1)
-
-    resp5 = request_to_json("post", "/environment/open")
-    assert resp5["success"] is True
-    assert wait_for_environment_to_be_created(10)
-
-    resp6 = request_to_json("get", "/status")
-    assert resp6["items_in_queue"] == 1
-    assert resp6["items_in_history"] == 0
-
-    resp7 = request_to_json("post", "/queue/start")
-    assert resp7["success"] is True
-
-    wait_for_queue_execution_to_complete(20)
-
-    resp8 = request_to_json("get", "/status")
-    assert resp8["items_in_queue"] == 0
-    assert resp8["items_in_history"] == 1
-
-    # Close the environment
-    resp9 = request_to_json("post", "/environment/close")
-    assert resp9 == {"success": True, "msg": ""}
-
-    wait_for_manager_state_idle(10)
-
-    import time as ttime
-
-    ttime.sleep(2)
-    resp10 = request_to_json("get", "/console_output", json={"nlines": 1000})
-    assert resp10["success"] is True
-    assert len(resp10["text"]) > 0
-    assert "RE Environment is ready" in resp10["text"], resp10["text"]
